third_party/blink/web_tests/platform/generic/fast/table/table-parts-in-inline-expected.txt

Test that a table part child of an inline is wrapped inside an inline-table.

This text should be on one line.
This text should be on one line.
This text should be on one line.
Testing cell:
PASS firstTop is lastTop
Testing row:
PASS firstTop is lastTop
Testing row-group:
PASS firstTop is lastTop
PASS successfullyParsed is true

TEST COMPLETE
